drivers/meson: rename vdec
