+++ drivers/media/platform/meson/vdec/codec_hevc.h
@@ -6,7 +6,7 @@
 #ifndef __MESON_VDEC_CODEC_HEVC_H_
 #define __MESON_VDEC_CODEC_HEVC_H_
 
-#include "vdec.h"
+#include "meson_vdec.h"
 
 extern struct amvdec_codec_ops codec_hevc_ops;
