+++ drivers/media/platform/meson/vdec/codec_mpeg4.h
@@ -6,7 +6,7 @@
 #ifndef __MESON_VDEC_CODEC_MPEG4_H_
 #define __MESON_VDEC_CODEC_MPEG4_H_
 
-#include "vdec.h"
+#include "meson_vdec.h"
 
 extern struct amvdec_codec_ops codec_mpeg4_ops;
