+++ drivers/media/platform/meson/vdec/codec_vp9.h
@@ -6,7 +6,7 @@
 #ifndef __MESON_VDEC_CODEC_VP9_H_
 #define __MESON_VDEC_CODEC_VP9_H_
 
-#include "vdec.h"
+#include "meson_vdec.h"
 
 extern struct amvdec_codec_ops codec_vp9_ops;
