+++ drivers/media/platform/meson/vdec/meson_vdec_ctrls.c
@@ -1,4 +1,4 @@
-#include "vdec_ctrls.h"
+#include "meson_vdec_ctrls.h"
 
 static int vdec_op_g_volatile_ctrl(struct v4l2_ctrl *ctrl)
 {
