+++ drivers/media/platform/meson/vdec/meson_vdec_ctrls.h
@@ -1,7 +1,7 @@
 #ifndef __MESON_VDEC_CTRLS_H_
 #define __MESON_VDEC_CTRLS_H_
 
-#include "vdec.h"
+#include "meson_vdec.h"
 
 int amvdec_init_ctrls(struct v4l2_ctrl_handler *ctrl_handler);
