+++ drivers/media/platform/meson/vdec/vdec_1.h
@@ -7,7 +7,7 @@
 #ifndef __MESON_VDEC_VDEC_1_H_
 #define __MESON_VDEC_VDEC_1_H_
 
-#include "vdec.h"
+#include "meson_vdec.h"
 
 extern struct amvdec_ops vdec_1_ops;
