+++ drivers/media/platform/meson/vdec/vdec_hevc.h
@@ -15,7 +15,7 @@
 #ifndef __MESON_VDEC_VDEC_HEVC_H_
 #define __MESON_VDEC_VDEC_HEVC_H_
 
-#include "vdec.h"
+#include "meson_vdec.h"
 
 extern struct amvdec_ops vdec_hevc_ops;
